+++ app/src/main/java/woowacourse/payments/ui/submitcard/SubmitCardScreenUiStateHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.addcard
+package woowacourse.payments.ui.submitcard
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
